refactor: positions hook to useSsovPositions
