+++ apps/dapp/src/components/ssov-new/Tables/Positions/BuyPositions.tsx
@@ -11,7 +11,7 @@ import {
 import format from 'date-fns/format';
 import Countdown from 'react-countdown';
 
-import { BuyPosition } from 'hooks/ssov/positions';
+import { BuyPosition } from 'hooks/ssov/useSsovPositions';
 import useVaultsData from 'hooks/ssov/useVaultsData';
 import useVaultStore from 'hooks/ssov/useVaultStore';
